+++ lib/app/models/user_settings.dart
@@ -58,6 +58,8 @@ class DefaultUserSettings {
   static const interruptionSkipBackward = Duration.zero;
   static const interruptionLongSkipThreshold = Duration(seconds: 30);
   static const interruptionLongSkipBackward = Duration.zero;
+  static const globalProgressEndLabel = ProgressEndLabel.total;
+  static const progressEndLabel = ProgressEndLabel.total;
 }
 
 @freezed
@@ -183,6 +185,12 @@ sealed class UserSettings with _$UserSettings {
 
     @Default(DefaultUserSettings.interruptionLongSkipBackward)
     Duration interruptionLongSkipBackward,
+
+    @Default(DefaultUserSettings.globalProgressEndLabel)
+    ProgressEndLabel globalProgressEndLabel,
+
+    @Default(DefaultUserSettings.progressEndLabel)
+    ProgressEndLabel progressEndLabel,
   }) = _UserSettings;
 
   factory fromJson(Map<String, dynamic> json) => _$UserSettingsFromJson(json);